requirements.txt
streamlit
python-dotenv
openai
supabase==2.0.3
pandas
